--- Project_AD/Level.cpp
@@ -1,6 +1,0 @@
-#include "Level.hpp"
-
-
-Level::Level() {
-
-}
