--- Project_AD/Level.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-class Level
-{
-public:
-	Level();
-
-
-
-private:
-
-};
-
