+++ Project_AD/Project_AD.vcxproj
@@ -137,7 +137,7 @@
   <ItemGroup>
     <ClCompile Include="Engine.cpp" />
     <ClCompile Include="Game.cpp" />
-    <ClCompile Include="Level.cpp" />
+    <ClCompile Include="LevelParser.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="tinyxml2.cpp" />
   </ItemGroup>
@@ -145,7 +145,6 @@
     <ClInclude Include="Engine.hpp" />
     <ClInclude Include="Button.hpp" />
     <ClInclude Include="Game.hpp" />
-    <ClInclude Include="Level.hpp" />
     <ClInclude Include="Level_buffer.hpp" />
     <ClInclude Include="LevelParser.hpp" />
     <ClInclude Include="Player.hpp" />
