+++ Project_AD/map2.tmx
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<map version="1.0" orientation="orthogonal" renderorder="right-down" width="60" height="34" tilewidth="32" tileheight="32" nextobjectid="25">
+<map version="1.0" orientation="orthogonal" renderorder="right-down" width="60" height="34" tilewidth="32" tileheight="32" nextobjectid="41">
  <tileset firstgid="1" name="Ground" tilewidth="32" tileheight="32" tilecount="256" columns="16">
   <image source="Resourses/Images/TXTilesetGround.png" width="512" height="512"/>
  </tileset>
@@ -41,9 +41,9 @@
 34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34,34
 </data>
  </layer>
- <objectgroup name="collision">
+ <objectgroup name="collision" offsetx="0" offsety="-1">
   <object id="1" name="solid" x="0" y="960" width="1922" height="126"/>
-  <object id="4" name="player" x="647" y="579" width="40" height="40"/>
+  <object id="4" name="player" x="104" y="917" width="40" height="40"/>
   <object id="7" name="solid" x="256" y="608" width="94.6667" height="30.6667"/>
   <object id="8" name="solid" x="128.667" y="351.333" width="94.6667" height="30.6667"/>
   <object id="10" name="solid" x="351.333" y="156.667" width="94.6667" height="30.6667"/>
@@ -53,7 +53,7 @@
   <object id="14" name="solid" x="1760.67" y="192.667" width="94.6667" height="30.6667"/>
   <object id="15" name="solid" x="1568.67" y="386" width="94.6667" height="30.6667"/>
   <object id="16" name="solid" x="1376.67" y="543.333" width="94.6667" height="30.6667"/>
-  <object id="17" name="solid" x="928.667" y="611.333" width="94.6667" height="30.6667"/>
+  <object id="17" name="solid" x="928.667" y="609.333" width="94.6667" height="30.6667"/>
   <object id="18" name="solid" x="608.667" y="767.333" width="94.6667" height="30.6667"/>
   <object id="19" name="solid" x="512.667" y="386" width="94.6667" height="30.6667"/>
   <object id="20" name="solid" x="1474" y="736.667" width="94.6667" height="30.6667"/>
